REFACTOR: allworkation 최적화
